the old _builtin_func no longer needs to be a decorator. Turn it into make_builtin_func, which slightly simplify some code
